backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
pydantic-settings==2.10.1
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5
pdfplumber==0.11.7
python-docx==1.2.0
keybert==0.9.0
sentence-transformers==5.1.0
numpy==2.3.2
scikit-learn==1.7.1
openai==1.99.9
reportlab==4.4.3
python-dateutil==2.9.0.post0
orjson==3.11.2
